Remove Heart Rate Sensor.
There are 2 main reasons I have put pretty much no effort into the MAX30101:
Getting good heart rate readings require ALOT of clever algorithms, there are some online, but tbh I have not found any that would work as good as you would expect. Good HR algorithms require fusion with accelerometer data etc. to be accurate when moving, as light may slip in below the watch and make the data "bad", hence accelerometer data needs to be fused toghether. So a lot of work would be needed, work I don't find fun :p
Mechanically it turned out to be a real challenge with the watch case. I have the sensor on the bottom of the PCB, and the sensor needs to sit directly touching the skin => bottom of watch needs to be super thin, while also having a connector for the dock => impossible (at least I could not make it work). Hence I actually only mounted the MAX30101 on one watch I built and not on the others as I could not get it mechanically to work.
I'm thinking for the next PCB version I do to drop the MAX30101 as it's just to challenging. It pretty much needs to sit on a separate PCB to make it mechanically work. Since it requires algorithms and makes casing hard, I'd much rather put something more useful in that space on the PCB.