Fix display glitching. Randomly turning white or going black.
Issue was that we would send display enter sleep, while we where at the same tie flushing data to it to display.
Fixed by making sure rendering finishes and giving the flushing thread some time to flush all data to the
display before we till the display to go to sleep.